Move regression tests to owning packages
